Moved class variable to separate file.  >> for variable
